% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATdisp.R
\name{ATDisp}
\alias{ATDisp}
\title{Compute and display the AT-disparity curve}
\usage{
ATDisp(sek)
}
\arguments{
\item{sek}{a vector with nucleotide sequence of genome, saved in DNAString}
}
\value{
returns the value of minimum of the AT-disparity curve and displays the plot of the same curve
}
\description{
A function, that computes the AT-disparity curve from Z-curve of the input genome and displays it consecutively.
}
\examples{
\dontrun{
sek<-readDNAStringSet('sequence.fasta')
sek<-sek[[1]]
ATDisp(sek)
}
}
\author{
Marko Lebo <markolebo48@gmai.com>
}

