% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumphase.R
\name{cumphase}
\alias{cumphase}
\title{Compute and display the cumulted phase signal of the input genome}
\usage{
cumphase(sek)
}
\arguments{
\item{sek}{a vector with nucleotide sequence of genome, saved in DNAString.}
}
\value{
Displays the cumulted phase signal of the input genome.
}
\description{
A function, that is used for computation and displaying the cumulted phase signal of the input genome
}
\examples{
\dontrun{
sek<-readDNAStringSet('sequence.fasta')
sek<-sek[[1]]
cumulphase <- cumphase(sek)
}
}
\author{
Marko Lebo <markolebo48@gmai.com>
}

