% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNA-walk.R
\name{Dnawalk}
\alias{Dnawalk}
\title{Compute and display the DNA-walk of the input genome}
\usage{
Dnawalk(sek)
}
\arguments{
\item{sek}{a vector with nucleotide sequence of genome, saved in DNAString.}
}
\value{
Displays the DNA-walk of the input genome.
}
\description{
A function, that is used for computation and displaying the DNA-walk of the input genome
}
\examples{
\dontrun{
sek<-readDNAStringSet('sequence.fasta')
sek<-sek[[1]]
DNAWalk <- Dnawalk(sek)
}
}
\author{
Marko Lebo <markolebo48@gmai.com>
}

