% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindOri.R
\name{FindOri}
\alias{FindOri}
\title{Detect the oriC region of a bacterial genome}
\usage{
FindOri(sek)
}
\arguments{
\item{sek}{a vector with nucleotide sequence of genome, saved in DNAString.}
}
\value{
returns a number of values in a \code{list}. In the first element of the \code{list} are parameters of the detected oriC region. The order of the parameteres are:

     \code{OriStart} - position genome, where the oriC region starts,

     \code{OriEnd} - position genome, where the oriC region starts,

     \code{OriLenght} - length of the detected oriC,

     \code{OriGC} - percentual GC-content of the oriC region,

     \code{OriAT} - percentual AT-content of the oriC region,

     \code{Min} - position of minimum of  GC-disparity curve,

     \code{DnaAcount} - number of detected DnaA-boxes in the detected oriC region.

     The second element of the returning \code{list} is a vector called \code{DistDnaA} in which the distances of the beginning of each
     detected DnaA-box from the beginning of oriC is returned.

     The last element of the returned \code{list} is the whole nucleotide sequence of the detected oriC region.

     This function also displays the GC-disparity curve of the input genome.
}
\description{
A function, that is used for oriC region detection in bacterial genomes
}
\examples{
\dontrun{
sek<-readDNAStringSet('sequence.fasta')
sek<-sek[[1]]
oriC <- FindOri(sek)
}
}
\author{
Marko Lebo <markolebo48@gmai.com>
}

