% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zcurve.R
\name{Zcurve}
\alias{Zcurve}
\title{Compute and display the GC-disparity curve}
\usage{
Zcurve(sek)
}
\arguments{
\item{sek}{a vector with nucleotide sequence of genome, saved in DNAString}
}
\value{
returns the value of minimum of the GC-disparity curve and displays the plot of the same curve
}
\description{
A function, that computes the GC-disparity curve from Z-curve of the input genome and displays it consecutively.
}
\examples{
\dontrun{
sek<-readDNAStringSet('sequence.fasta')
sek<-sek[[1]]
Min <- Zcurve(sek)
}
}
\author{
Marko Lebo <markolebo48@gmai.com>
}

